move getPayloadV5 boundary cases
